Fix bugs in training script

+++ src/binary_classifier/binary_classifier_region_abnormal.py
@@ -1,3 +1,4 @@
+import torch
 import torch.nn as nn
 
 
@@ -34,7 +35,7 @@ class BinaryClassifierRegionAbnormal(nn.Module):
         # only compute loss for logits that correspond to a class that was detected
         detected_logits = logits[class_detected]
         detected_region_has_sentence = region_is_abnormal[class_detected]
-        loss = self.loss_fn(detected_logits, detected_region_has_sentence)
+        loss = self.loss_fn(detected_logits, detected_region_has_sentence.type(torch.float32))
 
         if self.training:
             return loss
